+++ client/package.json
@@ -13,6 +13,7 @@
     "@types/react": "18.2.0",
     "@types/react-dom": "18.2.1",
     "autoprefixer": "10.4.14",
+    "axios": "^1.4.0",
     "next": "13.3.4",
     "postcss": "8.4.23",
     "react": "18.2.0",
